RocksDbUtility
